[org] Add org-appear support
Cherry-picked changes from @winsphinx. Didn't cherry-pick the attribution
because I think it's in the wrong place in the document, but I don't want to
correct it in case it gets merged where it currently is. I still hope the PR
will be merged and at that point I'll get the attribution by merging down.